019_DRF_PaginationFilterSearch/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
Faker==15.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
